Added user management in admin panel

+++ src/static/admin.js
@@ -11,11 +11,11 @@ const Book = ({
   return `
   <li class="card">
     <div class="form-group mx-2">
-      <label >Id:</label>
+      <label>Id:</label>
       <input type="text" class="form-control" disabled value="${id}" />
     </div>
     <div class="form-group mx-2">
-      <label >Título:</label>
+      <label>Título:</label>
       <input type="text" class="form-control" disabled value="${title}" />
     </div>
     <div class="form-group mx-2">
@@ -23,7 +23,7 @@ const Book = ({
       <input type="text" class="form-control" disabled value="${author}" />
     </div>
     <div class="form-group mx-2">
-      <label ption">Descripción:</label>
+      <label>Descripción:</label>
       <textarea class="form-control" rows="3" disabled>${description}</textarea>
     </div>
     <div class="form-group mx-2">
@@ -42,66 +42,59 @@ const Book = ({
   </li>
   `;
 };
+const User = ({ id, name, email, password }) => {
+  return `
+  <li class="card">
+    <div class="form-group mx-2">
+      <label>Id:</label>
+      <input type="text" class="form-control" disabled value="${id}" />
+    </div>
+    <div class="form-group mx-2">
+      <label >Nombre:</label>
+      <input type="text" class="form-control" disabled value="${name}" />
+    </div>
+    <div class="form-group mx-2">
+      <label>Email:</label>
+      <input type="text" class="form-control" disabled value="${email}" />
+    </div>
+    <div class="form-group mx-2">
+      <label>Contraseña:</label>
+      <textarea class="form-control" rows="3" disabled>${password}</textarea>
+    </div>
+    <button class="btn btn-danger m-2" onclick="userDeleteEv(${id})">Borrar usuario</button>
+  </li>
+  `;
+};
+
 // Validators
 const emptyAttributes = (book) => {
   return Object.keys(book).some((key) => book[key] === "");
 };
-
-// Requests
-const bookPost = async (book) => {
-  const myHeaders = new Headers();
-  if (window.auth_token) myHeaders.append("Authorization", window.auth_token);
-  myHeaders.append("Content-Type", "application/json");
-  const raw = JSON.stringify(book);
-  const requestOptions = {
-    method: "POST",
-    headers: myHeaders,
-    body: raw,
-    redirect: "follow",
-  };
-
-  try {
-    const response = await fetch("/api/book/", requestOptions);
-    await response.json();
-  } catch (error) {
-    alert("Failed book creation");
-  }
-};
-const bookGet = async () => {
-  const requestOptions = {
-    method: "GET",
-    redirect: "follow",
-  };
-  try {
-    const response = await fetch("/api/book/", requestOptions);
-    return await response.json();
-  } catch (error) {
-    console.log(error);
-    alert("No se pudo obtener los libros");
-    return [];
-  }
-};
-const bookDelete = async (id) => {
-  console.log("Delete", id);
-  const myHeaders = new Headers();
-  myHeaders.append("Authorization", window.auth_token);
-
+const logout = async () => {
+  if (!confirm("Estas seguro?")) return;
   const requestOptions = {
     method: "DELETE",
-    headers: myHeaders,
     redirect: "follow",
   };
-
-  const response = await fetch(`/api/book/${id}`, requestOptions);
-  if (response.ok) return;
-  alert("No se pudo eliminar el libro");
+  await fetch(`/login`, requestOptions);
+  window.location.replace("/admin-login");
 };
+
+// APIs
+const bookAPI = new ModelAPI("book", window.auth_token);
+const userAPI = new ModelAPI("user", window.auth_token);
 
 // Events
 const paintBooks = (books) => {
-  bookGet().then((bookList) => {
+  bookAPI.getModel().then((bookList) => {
     books.innerHTML = "";
     bookList.forEach((book) => (books.innerHTML += Book(book)));
+  });
+};
+const paintUsers = (users) => {
+  userAPI.getModel().then((userList) => {
+    users.innerHTML = "";
+    userList.forEach((user) => (users.innerHTML += User(user)));
   });
 };
 const bookSubmit = (ev, form) => {
@@ -114,28 +107,42 @@ const bookSubmit = (ev, form) => {
     alert("Faltan atributos en el formulario del libro");
     return;
   }
-  bookPost(json);
+  bookAPI.postModel(json);
   paintBooks(books);
+};
+const userSubmit = (ev, form) => {
+  ev.preventDefault();
+  const body = new FormData(form);
+  const json = Object.fromEntries(body);
+  if (emptyAttributes(json)) {
+    alert("Faltan atributos en el formulario del usuario");
+    return;
+  }
+  userAPI.postModel(json);
+  paintUsers(users);
 };
 const bookDeleteEv = (id) => {
   if (!confirm("Estas seguro?")) return;
-  bookDelete(id);
+  bookAPI.deleteModel(id);
   paintBooks(books);
 };
-const logout = async () => {
+const userDeleteEv = (id) => {
   if (!confirm("Estas seguro?")) return;
-  const requestOptions = {
-    method: "DELETE",
-    redirect: "follow",
-  };
-  await fetch(`/login`, requestOptions);
-  window.location.replace("/admin-login");
+  userAPI.deleteModel(id);
+  paintUsers(users);
 };
 
 const books = document.querySelector("#book-list");
+const users = document.querySelector("#user-list");
+
 const bookForm = document.querySelector("#book-form");
+const userForm = document.querySelector("#user-form");
+
 const logoutBtn = document.querySelector("#logout");
 
 bookForm.addEventListener("submit", (ev) => bookSubmit(ev, bookForm));
+userForm.addEventListener("submit", (ev) => userSubmit(ev, userForm));
 logoutBtn.addEventListener("click", logout);
+
 paintBooks(books);
+paintUsers(users);
